factorial of a number using function
The first line defines a function named factorial that takes one argument n (a number for which we want to calculate the factorial).
the if statement is for recursion. The factorial of Zero or One ,so if the n value is 0 or 1 then immediately it returns 1 without any further continuation. This is the recursive case.
If n is greater than 1,the function returns n*factorial(n-1).
it keeps calling itself with a smaller number until it reaches the base case.
the print statement prints the output